Moved compareLayout to a trait.
